requirements.txt
torch>=1.12.0
transformers>=4.20.0
scikit-learn>=1.1.0
pandas>=1.4.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pyyaml>=6.0
gdown>=4.6.0
plotly>=5.10.0
kaleido>=0.2.1